+++ app/Http/Requests/Booking/Admin/RouteAssign.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App\Http\Requests\Booking\Admin;
+
+use App\Http\Requests\Request;
+
+class RouteAssign extends Request
+{
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            'file' => 'required|file',
+        ];
+    }
+}