+++ CHANGELOG.md
@@ -6,10 +6,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.5.1](https://gitlab.com/daveroverts/Book-me-a-cookie/compare/v1.5.1...v1.5.0.c) - 2020-01-15
+
+### Added
+- Export for Multi-city event type (with e-mails and with CTOT's)
+- Added ``auth.isLoggedIn`` middleware to ``BookingController@edit()`` that you are logged in
+- E-mail template for Multi city event type
+- Route assign for multi-city event type
+
+### Changed
+- In the navbar, the Dutch VACC logo's are now routing to the Dutch VACC website
+- Route for `bookings.autoAssignForm` + ``bookings.autoAssign` is now URL friendly
+
 ## [v1.5.0c](https://gitlab.com/daveroverts/Book-me-a-cookie/compare/v1.5.0.c...v1.5.0.b) - 2020-01-07
 
 ### Changed
 - Fix for ``pic`` undefined in some views
+
+### Changed
+- ``event.admin.import`` view formats now include ``<abbr>``'s to better explain what's needed
 
 ## [v1.5.0b](https://gitlab.com/daveroverts/Book-me-a-cookie/compare/v1.5.0.b...v1.5.0.a) - 2020-01-04